Commented-out the old ASCII art logo of tripartiteOS to be replaced with a proper graphical version in the future
Commented-out the old ASCII art logo of tripartiteOS to be replaced with a proper graphical version in the future. The proper graphical version that the commented-out old ASCII art logo is going to be replaced with in the future is currently in progress of being made. In one of the following commits, the commented-out old ASCII logo of tripartiteOS will be replaced with a proper graphical version with the old ASCII art logo still being available in a separate file. The commented-out ASCII art logo of tripartiteOS that is going to be available in a separate file and that is going to be replaced with a proper graphical version will also be used on tripartiteOS's project's official web page. The tripartiteOS's project's official web page, apart from the commented-out ASCII art logo of tripartiteOS that is going to be available in a separate file and that is going to be replaced with a proper graphical version and that will also be used on the tripartiteOS's project's official web page, will also include a proper graphical version that the commented-out old ASCII art logo is going to be replaced with in the future is currently in progress of being made and that will be uploaded in one of the following commits.